LayoutTests/platform/mac/accessibility/aria-sort-expected.txt

button
column
row
This tests that aria-sort will be exposed correctly to the Mac accessibility API.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS accessibilityController.focusedElement.stringAttributeValue('AXSortDirection') is 'AXUnknownSortDirection'
PASS accessibilityController.focusedElement.stringAttributeValue('AXSortDirection') is 'AXAscendingSortDirection'
PASS accessibilityController.focusedElement.stringAttributeValue('AXSortDirection') is 'AXDescendingSortDirection'
PASS successfullyParsed is true

TEST COMPLETE
